pyproject.toml
[project]
name = "generate-facturx-py"
version = "1.3.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "factur-x>=3.15",
    "flask>=3.1.0",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.9",
]
